create channel factory and types
