+++ src/channels/PostMessage.ts
@@ -1,238 +1,217 @@
-import nacl from "tweetnacl";
-import util from "tweetnacl-util";
+import { Channel, ChannelFactory } from "./Channel.js";
+import {
+    ChannelClosedReceipt,
+    ChannelOpenedReceipt,
+    ChannelPrivateEvents,
+    ChannelPublicEvents,
+    CreateChannelArgs,
+    MessageSentReceipt,
+    SendMessageArgs
+} from "./types.js";
+import { randomNonce, timestamp } from "../utilities/index.js";
 
-class PostMessageChannel {
-  cid: string;
-  ctx: any; // todo - type this
-  origin: string;
-  target: Window;
-  publicKeys: any;
-  sharedKey: string;
-  onOpen: Function | undefined;
-  onClose: Function | undefined;
-  onMessage: Function | undefined;
-  closed: boolean = false
+const CONFIRMATION_TIMEOUT = 10000;
 
-  constructor({
-    cid,
-    origin,
-    target,
-    publicKeys,
-    sharedKey,
-    onOpen,
-    onMessage,
-    onClose,
-    encrypt,
-    decrypt,
-    sign,
-    verify,
-  }: {
-    cid: string,
-    origin: string,
-    target: Window,
-    publicKeys: any,
-    sharedKey?: string,
-    onOpen?: Function,
-    onMessage?: Function,
-    onClose?: Function,
-    encrypt: Function,
-    decrypt: Function,
-    sign: Function,
-    verify: Function
-  }) {
-    if (!cid || !origin || !target || !publicKeys || !sharedKey || !encrypt || !decrypt || !sign || !verify) {
-      throw new Error('missing required params')
+type PostMessageEvent = MessageEvent & {
+    data: {
+        cid?: string;
+        mid?: string;
+        type: string;
+        payload: any;
     }
-
-    this.ctx = { encrypt, decrypt, sign, verify }
-    this.cid = cid
-    this.origin = origin
-    this.target = target
-    this.publicKeys = publicKeys
-    this.sharedKey = sharedKey
-    this.onOpen = onOpen
-    this.onClose = onClose
-    this.closed = false
-
-    const messageHandler = onMessage ? (data, conn) => {
-      const { cid, mid, signature, ciphertext } = data;
-      const verified = this.ctx.verify({ data: ciphertext, signature, publicKey: this.publicKeys.signing });
-      if (!verified) return;
-      const message = this.ctx.decrypt({ data: ciphertext, sharedKey: this.sharedKey });
-      const signed = this.ctx.sign({ data: { cid, message } });
-      this.target.postMessage({ cid, mid, signature: signed, type: 'sparks-channel:message-confirmation' }, this.origin);
-      if (!this.closed && onMessage) onMessage(message, conn);
-    } : undefined
-
-    // todo -- either handle everything here or move it to the channel
-    // channels should hanve their own global listeners tbh
-    const handler = (event) => {
-      const { data, origin } = event;
-      const { cid, type } = data;
-      if (!cid || !type || !origin) return;
-      const isClosed = type === 'sparks-channel:closed';
-      const isMessage = type === 'sparks-channel:message';
-      if (isClosed) {
-        this.target.postMessage({ cid: this.cid, type: 'sparks-channel:closed-confirmation' }, this.origin);
-        if (onClose) onClose(cid, this);
-      } if (isMessage && messageHandler) {
-        messageHandler(data, this)
-      }
-    };
-
-    window.addEventListener('message', handler);
-    window.addEventListener('beforeunload', close);
-  }
-
-  async message(data) {
-    const mid = util.encodeBase64(nacl.randomBytes(16));
-    const ciphertext = this.ctx.encrypt({ data, sharedKey: this.sharedKey });
-    const signature = this.ctx.sign({ data: ciphertext, detached: true });
-    
-    return new Promise((resolve, reject) => {
-      const handler = (event: { data: any; source: any; origin: any; }) => {
-        const { data, source, origin } = event;
-        if (
-          data.mid === mid &&
-          source === this.target &&
-          origin === this.origin &&
-          data.type === 'sparks-channel:message-confirmation'
-        ) {
-          if (this.closed) return reject('channel closed');
-          else resolve(data.signature);
-          window.removeEventListener('message', handler);
-        }
-      };
-      this.target.postMessage({ cid: this.cid, mid: mid, type: 'sparks-channel:message', ciphertext, signature }, this.origin);
-      window.addEventListener('message', handler);
-    });
-  }
-
-  async close() {
-    return new Promise((resolve, reject) => {
-      const handleDisconnect = (event: { source: any; origin: string; data: any; }) => {
-        if (
-          event.source === this.target &&
-          event.origin === this.origin &&
-          event.data?.type === 'sparks-channel:closed-confirmation'
-        ) {
-          resolve(true);
-          window.removeEventListener('message', handleDisconnect);
-        }
-      };
-      this.target.postMessage({ cid: this.cid, type: 'sparks-channel:closed' }, this.origin);
-      window.addEventListener('message', handleDisconnect);
-      this.closed = true
-    });
-  }
 }
 
-class PostMessageManager {
-  channels: PostMessageChannel[];
-  constructor(ctx) {
-    this.channels = []
-    // inject dependencies
-    const open = this.open.bind(this)
-    this.open = args => {
-      open({
-        ...args,
-        publicKeys: ctx.publicKeys,
-        encrypt: ctx.encrypt.bind(ctx),
-        decrypt: ctx.decrypt.bind(ctx),
-        sign: ctx.sign.bind(ctx),
-        verify: ctx.verify.bind(ctx),
-        computeSharedKey: ctx.computeSharedKey.bind(ctx),
-      })
+export class PostMessageChannel extends Channel {
+    private callbacks: Map<ChannelPublicEvents | ChannelPrivateEvents, (args: any) => void> = new Map();
+
+    constructor(args) {
+        super(args);
     }
-  }
 
-  async close() {
-    this.channels.forEach((channel: { close: () => any; }) => channel.close())
-    this.channels = []
-  }
-
-  open(args) {
-    const { target, url } = args;
-    const { beforeOpen, onOpen, onClose, onMessage } = args;
-    const { computeSharedKey, publicKeys: ourPublicKeys } = args;
-    const { encrypt, decrypt, sign, verify } = args;
-    if (!computeSharedKey) throw new Error('computeSharedKey is required');
-    const channelId = util.encodeBase64(nacl.randomBytes(16))
-    let requestInterval
-
-    // let's setup a handler to manage connection request or confirmation
-    const handler = (event) => {
-      const { data: { cid, type, publicKeys }, origin, source } = event;
-      const confirming = type === 'sparks-channel:connection-confirmation';
-      const requesting = type === 'sparks-channel:connection-request';
-
-      if (confirming && !cid) return;                                                       // if we initiated and the channel id doesn't match return quietly
-      if (!type || !origin || !(confirming || requesting)) return;                          // if missing options return quietly, we can't control inbound postMessages
-      if (requesting && !source) throw new Error('Invalid source');                         // if they're requesting and there's no source throw
-      if (requesting && !publicKeys) throw new Error('Missing public keys');                // if they're requesting but there's missing a prive key throw
-      if (requesting && beforeOpen && !beforeOpen({ cid, origin, publicKeys })) return;     // let's process our beforeOpen callback to make sure it passes custom checks
-      const sharedKey = computeSharedKey({ publicKey: publicKeys.encryption });
-
-      // they're confirming we can setup the channel
-      if (confirming) {
-        if (!sharedKey) throw new Error('Failed to compute shared key');
-        // let's setup the channel if it doesn't exist
-        if (!this.channels.find(channel => channel.cid === cid)) {
-          const channel = new PostMessageChannel({ cid, origin, target: source, publicKeys, sharedKey, onOpen, onMessage, onClose, encrypt, decrypt, sign, verify });
-          this.channels.push(channel);
-          if (onOpen) onOpen(cid, channel);
+    /**
+     * open initiates the handshake to get the shared key established
+     */
+    public async open() {
+        const target = {
+            origin: window.location.origin,
+            publicKey: this.spark.controller.signingKeys.publicKey,
         }
 
-        clearInterval(requestInterval);
-        source.postMessage({ cid, type: 'sparks-channel:connection-confirmation', publicKeys: ourPublicKeys }, origin);
-        window.removeEventListener('message', handler);
+        this.target.window = window.open(this.target.origin);
 
-      } else if (requesting) {
-        if (!sharedKey) throw new Error('Failed to compute shared key');
-        // let's send a confirmation back to the source
-        source.postMessage({
-          type: 'sparks-channel:connection-confirmation',
-          cid,
-          publicKeys: ourPublicKeys,
-        }, origin);
+        return new Promise<ChannelOpenedReceipt | void>((resolve, reject) => {
+            this.send({ data: { type: ChannelPrivateEvents.OPEN_REQUEST, target }, confirm: true });
+            this.callbacks.set(ChannelPrivateEvents.OPEN_CONFIRMATION, resolve);
 
-        // setup the requester's channel first
-        const channel = new PostMessageChannel({ cid, origin, target: source, publicKeys, sharedKey, onOpen, onMessage, onClose, encrypt, decrypt, sign, verify });
-        this.channels.push(channel);
-        if (onOpen) onOpen(cid, channel);
-      }
+            setTimeout(() => { 
+                this.callbacks.clear();
+                reject({ message: 'channel opened this side but could not get confirmation' })
+            }, CONFIRMATION_TIMEOUT)
+        });
     }
-    window.addEventListener('message', handler);
 
-    // if has a target we can initiate the connection request
-    // if the target is a string open a window
-    if (!!url && typeof url === 'string') {
-      const targetOrigin = new URL(url).origin;
-      const targetWindow = target || window.open(url, '_blank');
-      if (!targetWindow || !targetOrigin) throw new Error('Failed to open window');
-      requestInterval = setInterval(() => {
-        targetWindow.postMessage({
-          type: 'sparks-channel:connection-request',
-          cid: channelId,
-          publicKeys: ourPublicKeys,
-        }, targetOrigin);
-      }, 1000);
+    /**
+     * this kills all channels and removes all listeners
+     */
+    public async close() {
+        return new Promise<ChannelClosedReceipt | void>(async (resolve, reject) => {
+            this.send({ data: { type: ChannelPrivateEvents.CLOSE_REQUEST }, confirm: true });
+            this.callbacks.clear();
+            this.callbacks.set(ChannelPrivateEvents.CLOSE_CONFIRMATION, resolve);
+
+            setTimeout(() => { 
+                this.callbacks.clear();
+                reject({ message: 'channel closed this side but could not get confirmation' })
+            }, CONFIRMATION_TIMEOUT)
+        });
     }
-  }
+
+    /**
+     * this sends a message to the other side
+     */
+    public async send(args: SendMessageArgs) {
+        const { confirm, data } = args;
+        return new Promise<MessageSentReceipt | void>((resolve, reject) => {
+            this.target.window.postMessage({
+                mid: randomNonce(16),
+                type: ChannelPublicEvents.MESSAGE,
+                data: data,
+                confirm: confirm,
+                timestamp: timestamp(),
+            }, this.target.origin);
+            this.callbacks.set(ChannelPrivateEvents.MESSAGE_CONFIRMATION, resolve);
+        });
+    }
+
+    public on(event: ChannelPublicEvents, callback: (args: any) => void): void {
+        this.callbacks.set(event, callback);
+    }
 }
 
-const PostMessage = Base => class PostMessage extends Base {
-  constructor(...args) {
-    super(...args)
-    this.postMessage = new PostMessageManager(this)
-  }
+export class PostMessage extends ChannelFactory {
+    private handler: (event: PostMessageEvent) => void;
+    private channels: Channel[] = [];
+
+    constructor(args) {
+        super(args);
+
+        const handleOpenRequest = (event: PostMessageEvent): any => {
+            // if there's no channel id, target or publicKey, return
+            // if the channel id already exists, return
+            // if the target url is not in the whitelist, return
+            // if the publicKey is not in the whitelist, return
+        };
+        const handleOpenConfirmation = (event: PostMessageEvent): any => { };
+        const handleCloseRequest = (event: PostMessageEvent): any => { };
+        const handleCloseConfirmation = (event: PostMessageEvent): any => { };
+        const handleMessage = (event: PostMessageEvent): any => { };
+        const handleMessageReceipt = (event: PostMessageEvent): any => { };
+        const handleOpen = (event: PostMessageEvent): any => { };
+        const handleClose = (event: PostMessageEvent): any => { };
+
+        this.handler = (event: PostMessageEvent): void => {
+            console.log(event)
+            const { data = {} } = event;
+            const { cid, mid, type, payload } = data || {};
+            if (!type || !payload || !(cid || mid)) {
+                return;
+            } else if (type === ChannelPrivateEvents.OPEN_REQUEST) {
+                handleOpenRequest(event);
+            } else if (type === ChannelPrivateEvents.OPEN_CONFIRMATION) {
+                handleOpenConfirmation(event);
+            } else if (type === ChannelPrivateEvents.CLOSE_REQUEST) {
+                handleCloseRequest(event);
+            } else if (type === ChannelPrivateEvents.CLOSE_CONFIRMATION) {
+                handleCloseConfirmation(event);
+            } else if (type === ChannelPrivateEvents.MESSAGE_CONFIRMATION) {
+                handleMessageReceipt(event);
+            } else if (type === ChannelPublicEvents.MESSAGE) {
+                handleMessage(event);
+            } else if (type === ChannelPublicEvents.OPEN) {
+                handleOpen(event);
+            } else if (type === ChannelPublicEvents.CLOSE) {
+                handleClose(event);
+            }
+        }
+
+        window.addEventListener('message', this.handler);
+        window.addEventListener('beforeunload', () => {
+            Promise.all(this.channels.map((channel) => channel.close()));
+        });
+    }
+
+    public create(args: CreateChannelArgs) {
+        const { request, receive } = args;
+        const validRequest = request && request.origin && request.publicKey;
+        const validReceive = request && typeof receive === 'function';
+        if (!validRequest && !validReceive) { 
+            throw new Error('invalid channel args, please provide a valid request object, or receive function');
+        }
+
+        const channel = new PostMessageChannel({ 
+            spark: this.spark, 
+            target: request,
+        });
+        // if request send open request
+        return channel;
+    }
 }
 
-PostMessage.type = 'channel'
-PostMessage.dependencies = {
-  encrypt: true,
-  hash: true,
-  sign: true,
+/**
+ const handleOpenRequest = (event: PostMessageEvent):any => {
+    // if there's no channel id, target or publicKey, return
+    // if the channel id already exists, return
+    // if the target url is not in the whitelist, return
+    // if the publicKey is not in the whitelist, return
+};
+
+const handleOpenConfirmation = (event: PostMessageEvent):any => {
+
+};
+
+const handleCloseRequest = (event: PostMessageEvent):any => {
+
+};
+
+const handleCloseConfirmation = (event: PostMessageEvent):any => {
+
+};
+
+const handleMessage = (event: PostMessageEvent):any => {
+
+};
+
+const handleMessageReceipt = (event: PostMessageEvent):any => {
+
+};
+
+const handleError = (event: PostMessageEvent):any => {
+
+};
+
+this.handler = (event: PostMessageEvent):any => {
+    const { data = {} } = event;
+    const { cid, mid, type, payload } = data || {};
+    if (!type || !payload || !(cid || mid)) return;
+    if (type === PostMessageEventTypes.OPEN_REQUEST) {
+        handleOpenRequest(event);
+    } else if (type === PostMessageEventTypes.OPEN_CONFIRMATION) {
+        handleOpenConfirmation(event);
+    } else if (type === PostMessageEventTypes.CLOSE_REQUEST) {
+        handleCloseRequest(event);
+    } else if (type === PostMessageEventTypes.CLOSE_CONFIRMATION) {
+        handleCloseConfirmation(event);
+    } else if (type === PostMessageEventTypes.MESSAGE) {
+        handleMessage(event);
+    } else if (type === PostMessageEventTypes.MESSAGE_RECEIPT) {
+        handleMessageReceipt(event);
+    } else if (type === PostMessageEventTypes.ERROR) {
+        handleError(event);
+    }
 }
 
-export default PostMessage
+window.addEventListener('message', this.handler);
+window.addEventListener('beforeunload', () =>{
+    Promise.all(this.channels.map((channel) => channel.close({ receipt: false })));
+});
+ */
