+++ src/channels/index.ts
@@ -1,1 +1,2 @@
 export * from './Channel.js'
+export * from './PostMessage.js'
